requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-js-asset==1.2.2
django-mptt==0.13.2
Faker==8.12.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
